IWF-105: Rename handleevent > event package
